:building_construction: Build [MVVM + Reposiory pattern] architecture.
Google公式が推奨する MVVM + Reposioryパターン アーキテクチャを適用
