+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/di/AppModule.kt
@@ -5,7 +5,7 @@ import io.ktor.client.engine.android.Android
 import io.ktor.client.plugins.HttpTimeout
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
 import io.ktor.serialization.kotlinx.json.json
-import jp.co.yumemi.android.codeCheck.repository.SearchRepositoryImpl
+import jp.co.yumemi.android.codeCheck.data.repository.SearchRepositoryImpl
 import kotlinx.serialization.json.Json
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
