+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/di/ViewModelModule.kt
@@ -1,7 +1,7 @@
 package jp.co.yumemi.android.codeCheck.di
 
-import jp.co.yumemi.android.codeCheck.SearchGitRepoViewModel
-import jp.co.yumemi.android.codeCheck.repository.SearchRepositoryImpl
+import jp.co.yumemi.android.codeCheck.app.SearchGitRepoViewModel
+import jp.co.yumemi.android.codeCheck.data.repository.SearchRepositoryImpl
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
