+++ app/src/main/res/layout/activity_top.xml
@@ -13,4 +13,4 @@
     app:layout_constraintRight_toRightOf="parent"
     app:layout_constraintTop_toTopOf="parent"
     app:navGraph="@navigation/nav_graph"
-    tools:context=".TopActivity" />
+    tools:context=".app.TopActivity" />
