+++ app/src/main/res/layout/fragment_search_git_repo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TopActivity">
+    tools:context=".app.TopActivity">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/searchBar"
