+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/searchGitRepoFragment"
-        android:name="jp.co.yumemi.android.codeCheck.SearchGitRepoFragment"
+        android:name="jp.co.yumemi.android.codeCheck.ui.search.SearchGitRepoFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_search_git_repo">
         <action
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/gitRepoDetailFragment"
-        android:name="jp.co.yumemi.android.codeCheck.GitRepoDetailFragment"
+        android:name="jp.co.yumemi.android.codeCheck.ui.detail.GitRepoDetailFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_git_repo_detail">
         <argument
             android:name="gitRepo"
-            app:argType="jp.co.yumemi.android.codeCheck.GitRepo" />
+            app:argType="jp.co.yumemi.android.codeCheck.data.model.GitRepo" />
     </fragment>
 
 </navigation>
